+++ Rivet/Rivet.psm1
@@ -40,6 +40,11 @@ class Rivet_Session
     [Rivet.Configuration.Database] $CurrentDatabase
 }
 
+enum Rivet_QueryKeyword
+{
+    Default = 1
+}
+
 $RivetSchemaName = 'rivet'
 $RivetMigrationsTableName = 'Migrations'
 $RivetMigrationsTableFullName = "[$($RivetSchemaName)].[$($RivetMigrationsTableName)]"
@@ -216,6 +221,7 @@ Test-RivetTypeDataMember -TypeName 'Rivet.Scale' -MemberName 'Value'
     Join-Path -Path $rivetModuleRoot -ChildPath 'Functions'
     Join-Path -Path $rivetModuleRoot -ChildPath 'Functions\Columns'
     Join-Path -Path $rivetModuleRoot -ChildPath 'Functions\Operations'
+    Join-Path -Path $rivetModuleRoot -ChildPath 'Functions\Metadata'
 } |
     Where-Object { Test-Path -Path $_ -PathType Container } |
     Get-ChildItem -Filter '*-*.ps1' |